requested changes pt2 --WIP
